// src/styles/_variables.scss.d.ts
export interface Variables {
  colorBackground: string;
  colorBackgroundDark: string;
  colorBackgroundLight: string;

  colorText: string;

  colorPrimary: string;
  colorPrimaryDark: string;
  colorPrimaryLight: string;

  colorSecondary: string;
  colorSecondaryDark: string;
  colorSecondaryLight: string;
}

export const styles: Variables;

export default styles;